python/requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
cryptography==41.0.7 